requirements.txt
django~=4.1.5
django-filter~=22.1
djangorestframework~=3.14.0
markdown~=3.4.1
pytest~=7.2.0
python-dotenv~=0.21.0
